location: replace `text` with end location
Most importantly, this provides an end location for all nodes, making
it much easier (and less expensive) to highlight issues more precisely,
from the start location to the end.

Naturally, not having `text` embedded in the AST also reduces its size.

While the performance benefit of this is yet unknown, it does at least
save space. The size of the serialized AST for Regal goes from ~12 MB
to 10.

Also in this PR:
- Create a separate encoder for `ast.Head`, so that we may marshal (and
  test) it without having a full `ast.Rule`.

Signed-off-by: Anders Eknert <anders@eknert.com>